+++ src/models/paymentOrders/paymentOrder.ts
@@ -6,9 +6,9 @@ import { Language } from "../language";
 import { MetaData } from "../generics/MetaData-resource";
 import { State } from "../enums/state";
 import { Urls } from "../generics/urls";
-import { CurrentPaymentResponse } from "./currentPaymentResponse";
+import { CurrentPaymentResponse } from "./response models/currentPaymentResponse";
 import { OrderItemListResponse } from "./orderItemListResponse";
-import { PayerResponse } from "./payerResponse";
+import { PayerResponse } from "./response models/payerResponse";
 
 export interface PaymentOrder extends Identifiable {
     amount: Amount;
